Created TopicSelect component
